splitlunch: Fix transactions where user is uninvolved
This inverts the logic for `self_paid` in both `_get_expense_impact` and
`_get_payment_impact`, defaulting it to False. Previously, if the user
was completely uninvolved in a transaction (e.g. it was solely between
other members of a group), `self_paid` would be incorrectly set to True.
Additionally, filter these transactions out of the refresh sync.

And finally, in order to make the logic easier to test, the functions
are pulled to the top-level of the `lunchmoney_splitwise` module, so
that they can be tested without requiring the full machinery of the
SplitLunch class.