splitlunch: Rework to not require personal_share
The personal_share field worked for the subset of transactions that are
currently eligible for the "splitlunch refresh" sync function, but it
would not be correct if we expanded refresh to also allow syncing
self-paid and payment transactions. Meanwhile, the only other usage of
that field was for the split in SplitLunchImport transactions, which
were always self-paid non-payments. It's easy to switch SplitLunchImport
to use financial_impact directly, and in the case of refresh sync, it's
correct for all transactions (since financial_impact represents the
change to the user's balance in Splitwise).

For simplicity, this also flips the sign of the financial_impact field.
A positive financial_impact now indicates that the user owes money to
other Splitwise users as a result of the transaction. (The old logic
effectively did this in the cases it cared about by passing most values
to `abs`.)